% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{areaDivide}
\alias{areaDivide}
\title{Function to divide up areas. Input grid areas, returns East vs. West}
\usage{
areaDivide(area)
}
\description{
Function to divide up areas. Input grid areas, returns East vs. West
}
\keyword{internal}
