% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{areafunc}
\alias{areafunc}
\title{Stratum designations from Scott-Denton paper, and shrimp observer manual for GOM shrimp areas}
\usage{
areafunc(x)
}
\description{
Stratum designations from Scott-Denton paper, and shrimp observer manual for GOM shrimp areas
}
\keyword{internal}
