% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{areaGOM}
\alias{areaGOM}
\title{Function to convert new areas to old areas from Kevin McCarthy}
\usage{
areaGOM(x)
}
\description{
Function to convert new areas to old areas from Kevin McCarthy
}
\keyword{internal}
