% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BycatchEstimator.R
\docType{package}
\name{BycatchEstimator-package}
\alias{BycatchEstimator}
\alias{BycatchEstimator-package}
\title{Bycatch Estimator}
\description{
Generic Model-Based Bycatch Estimation Procedure
}
\details{
The code can estimate both total bycatch, calculated by expanding a sample, such as an observer database, to total effort from logbooks or landings records, and an annual index of abundance, calculated only from the observer data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ebabcock.github.io/BycatchEstimator/}
}

}
\author{
\strong{Maintainer}: Elizabeth Babcock \email{ebabcock@rsmas.miami.edu}

Authors:
\itemize{
  \item Elizabeth Babcock \email{ebabcock@rsmas.miami.edu}
  \item William Harford \email{bill@natureanalytics.ca} [contributor]
  \item Ana Adao \email{aadao@natureanalytics.ca} [contributor]
}

}
\keyword{internal}
