% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetup.R
\name{bycatchSetup}
\alias{bycatchSetup}
\title{Bycatch estimation data setup}
\usage{
bycatchSetup(
  obsdat,
  logdat,
  yearVar,
  latitudeVar = NULL,
  longitudeVar = NULL,
  spatialBin = 1,
  obsEffort,
  logEffort,
  obsCatch,
  catchUnit,
  catchType,
  logNum = NA,
  sampleUnit,
  factorVariables,
  numericVariables,
  EstimateBycatch = TRUE,
  baseDir = getwd(),
  runName,
  runDescription,
  common,
  sp,
  reportType = "html"
)
}
\arguments{
\item{obsdat}{Observer data set}

\item{logdat}{Logbook data set}

\item{yearVar}{Character. The column name of the year variable in \code{obsdat} and \code{logdat}. Both input files must contain the same variable name for year.}

\item{latitudeVar}{Optional name of latitude column, to get spatial plots.}

\item{longitudeVar}{Optional name of longitude column, to get spatial plots.}

\item{spatialBin}{Optional. If latitude and longitude are provided, this is the size of the spatial bin in degrees to use for spatial plots. Default is 1 degree.}

\item{obsEffort}{Character. The column name of the effort variable in \code{obsdat}. This variable must have the same effort units as \code{logEffort}}

\item{logEffort}{Character. The column name of the effort variable in \code{logdat}. Optional and only used when estimating bycatch. This variable must have the same effort units as \code{obsEffort}}

\item{obsCatch}{Character vector. The name of the column(s) in \code{obsdat} that contain catch. If it is a vector, order of variable names must follow the same order as names provided in \code{common} and \code{sp}}

\item{catchUnit}{Character vector. Give units of catch (e.g., number) to go in plot labels. Must be a vector of the same length as \code{sp}}

\item{catchType}{Character vector. Give type of catch (e.g., dead discards) to go in plot labels. Must be a vector of the same length as \code{sp}}

\item{logNum}{Character vector. The name of the column in \code{logdat} that gives the number of sample units (e.g., trips or sets). If the logbook data is not aggregated (i.e. each row is a sample unit) set value to NA}

\item{sampleUnit}{Character. What is the sample unit in \code{logdat}? e.g. sets or trips.}

\item{factorVariables}{Character vector. Specify which variables should be interpreted as categorical, ensuring factor format on these variables. These variables must have identical names and factor levels in \code{obsdat} and \code{logdat}}

\item{numericVariables}{Character vector. Specify which variables should be interpreted as numeric. These variables must have identical names in \code{obsdat} and \code{logdat}. If there are no numeric variables, set numericVariables=NA.}

\item{EstimateBycatch}{Logical. Defaults to TRUE. If TRUE, you must provide logbook data or some other source of total effort to \code{logdat}. FALSE will produced data summaries of \code{obsdat} only.}

\item{baseDir}{Character. A directory to save output. Defaults to current working directory.}

\item{runName}{Characer. Give a name to the run, which will be used to set up a directory for the outputs}

\item{runDescription}{Character. Brief summary of the run, which will be used to set up a directory for the outputs}

\item{common}{Character vector. Provide a common name for the species used in output filess. Can be a vector of names to do multiple species at the same time.}

\item{sp}{Character vector. Provide a scientific name for the species used in output files. Can be a vector of names to do multiple species at the same time}

\item{reportType}{Character. Choose type of report to be produced. Options are html (default), pdf or both.}
}
\description{
Sets global conditions, makes a preliminary data summary and data checks (make plots and tables of the data).
}
\examples{
\dontrun{
library(BycatchEstimator)
setupObj<-bycatchSetup(
obsdat = obsdatExample,
logdat = logdatExample,
yearVar = "Year",
latitudeVar = "Latitude",
longitudeVar = "Longitude",
spatialBin = 1,
obsEffort = "sampled.sets",
logEffort = "sets",
obsCatch = "Catch",
catchUnit = "number",
catchType = "dead discard",
logNum = NA,
sampleUnit = "trips",
factorVariables = c("Year","season"),
numericVariables = NA,
EstimateBycatch = TRUE,
baseDir = getwd(),
runName = "SimulatedExample",
runDescription = "Example with simulated data",
common = "Simulated species",
sp = "Genus species",
reportType = "html"
)}
}
\keyword{Fitting}
\keyword{functions}
