% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetup.R
\name{bycatchSetup_new}
\alias{bycatchSetup_new}
\title{Bycatch estimation data setup (new)}
\usage{
bycatchSetup_new(
  obsdat,
  logdat,
  yearVar,
  obsEffort,
  logEffort,
  obsCatch,
  catchUnit,
  catchType,
  logNum,
  sampleUnit,
  factorVariables,
  numericVariables,
  logUnsampledEffort = NULL,
  includeObsCatch = FALSE,
  matchColumn = NA,
  EstimateIndex = FALSE,
  EstimateBycatch = TRUE,
  baseDir = getwd(),
  runName,
  runDescription,
  common,
  sp,
  reportType
)
}
\arguments{
\item{obsdat}{Observer data set}

\item{logdat}{Logbook data set}

\item{yearVar}{Character. The column name of the year variable in \code{obsdat} and \code{logdat}. Both input files must contain the same variable name for year.}

\item{obsEffort}{Character. The column name of the effort variable in \code{obsdat}. This variable must have the same effort units as \code{logEffort}}

\item{logEffort}{Character. The column name of the effort variable in \code{logdat}. Optional and only used when estimating bycatch. This variable must have the same effort units as \code{obsEffort}}

\item{obsCatch}{Character vector. The name of the column(s) in \code{obsdat} that contain catch. If it is a vector, order of variable names must follow the same order as names provided in \code{common} and \code{sp}}

\item{catchUnit}{Character vector. Give units of catch (e.g., number) to go in plot labels. Must be a vector of the same length as \code{sp}}

\item{catchType}{Character vector. Give type of catch (e.g., dead discards) to go in plot labels. Must be a vector of the same length as \code{sp}}

\item{logNum}{Character vector. The name of the column in \code{logdat} that gives the number of sample units (e.g., trips or sets). If the logbook data is not aggregated (i.e. each row is a sample unit) set value to NA}

\item{sampleUnit}{Character. What is the sample unit in \code{logdat}? e.g. sets or trips.}

\item{factorVariables}{Character vector. Specify which variables should be interpreted as categorical, ensuring imposes factor format on these variables. These variables must have identical names and factor levels in \code{obsdat} and \code{logdat}}

\item{numericVariables}{Character vector. Specify which variables should be interpreted as numeric. These variables must have identical names in \code{obsdat} and \code{logdat}}

\item{logUnsampledEffort}{Character. The name of the unsampled effort variable in in \code{logdat}. Optional and used to specify a column for effort that is not sampled, in trips with observers. This can be zero in all cases if observers sample 100\% of effort in sampled trips. Only used when \code{includeObsCatch} is TRUE}

\item{includeObsCatch}{Logical. Set to TRUE if (1) the observed sample units can be matched to the logbook sample units and (2) you want to calculate total bycatch as the observed bycatch plus the predicted unobserved bycatch. This doesn't work with aggregated logbook effort.}

\item{matchColumn}{Character. If \code{includeObsCatch} is TRUE, give the name of the column that matches sample units between the observer and logbook data. Otherwise, this can be NA}

\item{EstimateIndex}{Logical. What would you like to estimate? You may calculate either an annual abundance index, or total bycatch, or both.}

\item{EstimateBycatch}{Logical. What would you like to estimate? You may calculate either an annual abundance index, or total bycatch, or both. If you want total bycatch, you must provide logbook data or some other source of total effort to \code{logdat}.}

\item{baseDir}{Character. A directory to save output. Defaults to current working directory.}

\item{runName}{Characer. Give a name to the run, which will be used to set up a directory for the outputs}

\item{runDescription}{Character. Brief summary of the run, which will be used to set up a directory for the outputs}

\item{common}{Character vector. Provide a common name for the species used in bycatch and index estimation. Can be a vector of names to do multiple species at the same time.}

\item{sp}{Character vector. Provide a scientific name for the species used in bycatch and index estimation. Can be a vector of names to do multiple species at the same time}

\item{reportType}{Character. Choose type of data checks report to be produced. Options are pdf, html or both.}
}
\description{
Sets global conditions, makes a preliminary data summary and data checks (make plots and tables of the data).
}
\examples{
\dontrun{
library(BycatchEstimator)
setupObj<-bycatchSetup_new( #do we need to include optional arguments in the example?
obsdat = obsdatExample,
logdat = logdatExample,
yearVar = "Year",
obsEffort = "sampled.sets",
logEffort = "sets",
obsCatch = "Catch",
catchUnit = "number",
catchType = "dead discard"
logNum = NA,
sampleUnit = "trips",
factorVariables = c("Year","season"),
numericVariables = NA,
logUnsampledEffort = NULL,
includeObsCatch  = FALSE,
matchColumn = NA,
EstimateIndex = FALSE,
EstimateBycatch = TRUE,
baseDir = getwd(),
runName = "SimulatedExample",
runDescription = "Example with simulated data",
common = "Simulated species",
sp = "Genus species",
reportType = "html"
)}
}
\keyword{Fitting}
\keyword{functions}
