% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{CheckForPositivesPlot}
\alias{CheckForPositivesPlot}
\title{Same for plotting}
\usage{
CheckForPositivesPlot(datval, species, variables)
}
\arguments{
\item{datval}{Value}

\item{species}{Value}

\item{variables}{Value}
}
\description{
Same for plotting
}
