% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCheck.R
\name{dataCheck}
\alias{dataCheck}
\title{Make plots and tables of the data}
\usage{
dataCheck(setupObj)
}
\arguments{
\item{setupObj}{An object produced by \code{bycatchSetup}.}
}
\description{
Make plots and tables of the data
}
\examples{
\dontrun{
library(BycatchEstimator)
dataCheck(setupObj)
#' }
#-------------------------------------------------
}
\keyword{Data}
\keyword{function}
\keyword{plotting}
