% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{deltaEstimatorMean}
\alias{deltaEstimatorMean}
\title{Function for Pennington(1983) method mean}
\usage{
deltaEstimatorMean(x)
}
\arguments{
\item{x}{is vector of data input}
}
\description{
Output is mean of the delta estimator
}
\keyword{internal}
