% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{deltaEstimatorSE2}
\alias{deltaEstimatorSE2}
\title{Function for Pennington(1983) method SE of the mean sqaured}
\usage{
deltaEstimatorSE2(x)
}
\arguments{
\item{x}{is vector of data input}
}
\description{
Output is SE squared of the delta estimator
}
\keyword{internal}
