% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{deltaEstimatorVar}
\alias{deltaEstimatorVar}
\title{Function for Pennington(1983) method variance}
\usage{
deltaEstimatorVar(x)
}
\arguments{
\item{x}{is vector of data input}
}
\description{
Output is variance of the delta estimator
}
\keyword{internal}
