% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{fishTimeFunc}
\alias{fishTimeFunc}
\title{Function to calculate number of hours fished in each day counted as first set to last haul}
\usage{
fishTimeFunc(timeout, timein, prop.sampled = 1)

fishTimeFunc(timeout, timein, prop.sampled = 1)
}
\description{
Function to calculate number of hours fished in each day counted as first set to last haul

fishTimeFunc
}
\keyword{internal}
