% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{FitModelFunc}
\alias{FitModelFunc}
\title{Function to fit a specified model formula and print outputs}
\usage{
FitModelFunc(formula1, formula2, modType, obsdatval, outputDir)
}
\arguments{
\item{formula1}{Value}

\item{formula2}{Value}

\item{modType}{Value}

\item{obsdatval}{Value}

\item{outputDir}{Value}
}
\description{
Function to fit a specified model formula and print outputs
}
