% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{FitModelFuncCV}
\alias{FitModelFuncCV}
\title{Function to fit a specified model formula and print outputs for Cross validation}
\usage{
FitModelFuncCV(formula1, modType, obsdatval)
}
\arguments{
\item{formula1}{Value}

\item{modType}{Value}

\item{obsdatval}{Value}
}
\description{
Function to fit a specified model formula and print outputs for Cross validation
}
\keyword{internal}
