% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{get_fixed_model_matrix}
\alias{get_fixed_model_matrix}
\title{Extract fixed effect model matrix columns corresponding to the vcov matrix}
\usage{
get_fixed_model_matrix(fit, data, is_TMB)
}
\description{
Extract fixed effect model matrix columns corresponding to the vcov matrix
}
\keyword{internal}
