% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{getdatefunc}
\alias{getdatefunc}
\title{Function to convert data in excel format with date and time separated by a blank into an R format date}
\usage{
getdatefunc(x, dateformat = "\%m/\%d/\%Y")
}
\arguments{
\item{x}{data}

\item{dateformat}{Date format, default is \%m/\%d/\%Y.}
}
\description{
Function to convert data in excel format with date and time separated by a blank into an R format date
}
\keyword{internal}
