% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{getdatefunc2}
\alias{getdatefunc2}
\title{Function 2 to convert data in excel format with date and time separated by a blank into an R format date}
\usage{
getdatefunc2(x, dateformat = "\%d\%b\%Y")
}
\arguments{
\item{x}{data}

\item{dateformat}{Date format, default is \%d\%b\%Y.}
}
\description{
Function 2 to convert data in excel format with date and time separated by a blank into an R format date
}
