% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{getDesignEstimatesUngroup}
\alias{getDesignEstimatesUngroup}
\title{Allocate design estimates to grouping variable not in the designVars}
\usage{
getDesignEstimatesUngroup(logdatval, groupVar, designVars, designstratadf)
}
\description{
Allocate design estimates to grouping variable not in the designVars
}
\keyword{internal}
