% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{getModelSummaryTable}
\alias{getModelSummaryTable}
\title{Return a table of model parameters and summaries}
\usage{
getModelSummaryTable(modfits, modTypes)
}
\arguments{
\item{modfits}{A list of fitted model objects}

\item{modTypes}{A corresponding vector of the model types as specified in modelTry}
}
\description{
Return a table of model parameters and summaries
}
\keyword{internal}
