% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{getPooling}
\alias{getPooling}
\title{Function to setup pooling if requested for design estimators}
\usage{
getPooling(
  obsdatval,
  logdatval,
  minStrataUnit,
  designVars,
  pooledVar,
  poolTypes,
  adjacentNum
)
}
\arguments{
\item{obsdatval}{Value}

\item{logdatval}{Value}

\item{minStrataUnit}{Value}

\item{designVars}{Value}

\item{pooledVar}{Value}

\item{poolTypes}{Value}

\item{adjacentNum}{Value}
}
\description{
Function to setup pooling if requested for design estimators
}
\keyword{internal}
