% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{getSimDeltaLN}
\alias{getSimDeltaLN}
\title{Generate mean and standard error of predictions for delta lognormal by simulation}
\usage{
getSimDeltaLN(modfitBin, modfitLnorm, df1, nsim = 10000)
}
\arguments{
\item{modfitBin}{Value}

\item{modfitLnorm}{Value}

\item{df1}{Value}

\item{nsim}{Value}
}
\description{
Generate mean and standard error of predictions for delta lognormal by simulation
}
\keyword{internal}
