% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{getSimSE}
\alias{getSimSE}
\title{Generate standard errors of predictions from simulation from regression coefficients and their var/covar matrix}
\usage{
getSimSE(modfit, df1, transFunc = "none", offsetval = NULL, nsim)
}
\arguments{
\item{modfit}{Value}

\item{df1}{Value}

\item{transFunc}{Value}

\item{offsetval}{Value}

\item{nsim}{Value}
}
\description{
Generate standard errors of predictions from simulation from regression coefficients and their var/covar matrix
}
