% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{Gm}
\alias{Gm}
\title{Function for Pennington(1983) method}
\usage{
Gm(m, t, jmax = 10)
}
\arguments{
\item{m}{m is the number of non-zero values}

\item{t}{t is a real number input}

\item{jmax}{jmax is the upper limit of an infinute sum used in calculations. 10 is usually sufficient}
}
\description{
Output is result of Gm(t) function for use in calculations of the design based delta estimator
}
\keyword{internal}
