% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{goodman.var}
\alias{goodman.var}
\title{Exact variance of the product of two independent variables, from Goodman (1960)}
\usage{
goodman.var(x, y)
}
\description{
Exact variance of the product of two independent variables, from Goodman (1960)
}
\keyword{internal}
