% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{length.unique}
\alias{length.unique}
\title{Function to count the number of unique levels in a vector}
\usage{
\method{length}{unique}(x)
}
\description{
Function to count the number of unique levels in a vector
}
\keyword{internal}
