% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExamples.R
\docType{data}
\name{LLSIM_BUM_Example_logbook}
\alias{LLSIM_BUM_Example_logbook}
\title{LLSIM-based example logbook program data set}
\format{
A tibble with 25 columns. Each row is a fishing trip.
\describe{
\item{trip}{Trip idenifier}
\item{Year}{Year}
\item{month}{month of the year}
\item{gear}{Gear code}
\item{light}{Light code}
\item{fleet}{Fleet number}
\item{bait}{Bait code}
\item{hook}{Hook code}
\item{hooks}{Effort variable, hooks}
\item{sets}{Number of sets in the trip}
\item{SWO}{Swordfish catch in numbers}
\item{BUM}{Blue marlin catch in numbers}
\item{lat5}{Latitude assigned to 5 degree grid}
\item{lon5}{Longitude assigned to 5 degree grid}
\item{lat}{Trip latitude}
\item{lon}{Trip longitude}
\item{hbf}{Hooks between floats}
\item{habSWO}{SDM habitat variable for swordfish}
\item{habBUM}{SDM habitat variables for blue marlin}
\item{season}{Categorical season variable}
\item{area}{Categorical season variable}
\item{unsampledEffort}{Unsampled effort}
}
}
\source{
Simulated data
}
\usage{
LLSIM_BUM_Example_logbook
}
\description{
LLSIm BUM by trip, with 5\% observer coverage including observed catch
}
\keyword{LLSIM}
\keyword{Simulated}
\keyword{data}
\keyword{from}
\keyword{set}
