% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BycatchEstimator.R
\docType{data}
\name{LLSIM_BUM_Example_observer}
\alias{LLSIM_BUM_Example_observer}
\title{LLSIM-based example observer program data set}
\format{
A tibble with 22 columns. Each row is a fishing trip.
\describe{
\item{Year}{Year}
\item{month}{month of the year}
\item{gear}{Gear code}
\item{light}{Light code}
\item{fleet}{Fleet number}
\item{bait}{Bait code}
\item{hook}
\item{hooks}
\item{sets}
\item{SWO}
\item{BUM}
\item{lat5}
\item{lon5}
\item{lat}
\item{lon}
\item{hbf}
\item{habSWO}
\item{habBUM}
\item{season}
\item{area}
}
}
\source{
Simulated data
}
\usage{
LLSIM_BUM_Example_observer
}
\description{
LLSIm BUM by trip, with 5\% observer coverage including observed catch in totals April 17 2022
}
\keyword{datasets}
