% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{lnorm.mean}
\alias{lnorm.mean}
\title{Calculate lognormal mean and standard error from normal mean and se}
\usage{
lnorm.mean(x1, x1e)
}
\description{
Calculate lognormal mean and standard error from normal mean and se
}
\keyword{internal}
