% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{lo.se}
\alias{lo.se}
\title{Variance of a product from Lo et al. (1992)}
\usage{
lo.se(x1, x1e, x2, x2e)
}
\arguments{
\item{x1}{Value}

\item{x1e}{Value}

\item{x2}{Value}

\item{x2e}{Value}
}
\description{
Variance of a product from Lo et al. (1992)
}
