% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOutputs.r
\name{loadOutputs}
\alias{loadOutputs}
\title{loadOutputs}
\usage{
loadOutputs(
  baseDir = getwd(),
  runName,
  runDate = Sys.Date(),
  designScenarios = NULL,
  modelScenarios = NULL
)
}
\arguments{
\item{baseDir}{The base directory for the runs, same as bycatchSetup.}

\item{runName}{The run name, same as bycatchSetup.}

\item{runDate}{The date when the model was run. Defaults to current date, but can be set to read in models previously run.}

\item{designScenarios}{Character vector of designScenario values from original run. NULL to read in no design-based results.}

\item{modelScenarios}{Character vector of modelScenario values from original run.}
}
\value{
Returns a list with the setupObj from the specified run, a list called designObjList
which contains the design-based model inputs and outputs for each designScenario, modelObjList,
which is the same for the models in modelScenarios, a data frame called allYearEstimates which
is the annual estimates across all design-based and model-based scenarios in a format suitable
for ggplot, allGroupEstimates, which is the same by the groups defined in groupVar
and predictionGroups, as well as the results by groups for models and design-based estimators
separately (allModResultsGroup, allDesignResultsGroup).
}
\description{
Reads in all the R objects created during runs of the bycatchEstimator for further analysis
}
\keyword{outputs}
\keyword{reload}
