% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExamples.R
\docType{data}
\name{logdatExample}
\alias{logdatExample}
\title{Logbook simple example data set}
\format{
A tibble with 5 columns.
\describe{
\item{EW}{East/West spatial grouping variable}
\item{season}{season temporal grouping variable}
\item{Year}{Year variable}
\item{sets}{The effort variable e.g., e.g. 1000 hook hours}
\item{trips}{Number of sample units in the record, trips in this example. In this example logbook data is not aggregated, thus each row is 1 trip}
}
}
\source{
Simulated data
}
\usage{
logdatExample
}
\description{
Example data set for logbook. Can be used to run bycatch estimation procedure along with observer example data set.
}
\keyword{Simple}
\keyword{data}
\keyword{set}
\keyword{simulated}
