% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{makeIndexVar}
\alias{makeIndexVar}
\title{Function to get an abundance index with SE}
\usage{
makeIndexVar(
  modfit1,
  modfit2 = NULL,
  modType,
  newdat,
  nsims,
  printOutput = FALSE,
  catchType = NULL,
  common = NULL,
  dirname = NULL,
  run = NULL
)
}
\arguments{
\item{modfit1}{Value}

\item{modfit2}{Value}

\item{modType}{Value}

\item{newdat}{Value}

\item{nsims}{Value}

\item{printOutput}{Value}

\item{catchType}{Value}

\item{common}{Value}

\item{dirname}{Value}

\item{run}{Value}
}
\description{
Function to get an abundance index with SE
}
\keyword{internal}
