% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{makePredictions}
\alias{makePredictions}
\title{Function to predict CPUE without variances to get predictions quickly for cross validation}
\usage{
makePredictions(modfit1, modfit2 = NULL, modType, newdat, obsdatval = NULL)
}
\arguments{
\item{modfit1}{Value}

\item{modfit2}{Value}

\item{modType}{Value}

\item{newdat}{Value}

\item{obsdatval}{Value}
}
\description{
Function to predict CPUE without variances to get predictions quickly for cross validation
}
\keyword{internal}
