% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{makePredictionsSimVar}
\alias{makePredictionsSimVar}
\title{Generate standard errors and confidence intervals of predictions from simulation from regression coefficients and their var/covar matrix}
\usage{
makePredictionsSimVar(
  modfit1,
  modfit2 = NULL,
  modtype,
  newdat,
  obsdatval = NULL,
  includeObsCatch,
  nsim,
  requiredVarNames,
  CIval,
  printOutput = TRUE,
  catchType,
  common,
  dirname,
  run
)
}
\arguments{
\item{modfit1}{Value}

\item{modfit2}{Value}

\item{modtype}{Value}

\item{newdat}{Value}

\item{obsdatval}{Value}

\item{includeObsCatch}{Value}

\item{nsim}{Value}

\item{requiredVarNames}{Value}

\item{CIval}{Value}

\item{printOutput}{Value}

\item{catchType}{Value}

\item{common}{Value}

\item{dirname}{Value}

\item{run}{Value}
}
\description{
Generate standard errors and confidence intervals of predictions from simulation from regression coefficients and their var/covar matrix
}
