% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{MakeSummary}
\alias{MakeSummary}
\title{Function to make data summarizes including ratio estimate at}
\usage{
MakeSummary(obsdatval, logdatval, strataVars, EstimateBycatch)
}
\arguments{
\item{obsdatval}{Value}

\item{logdatval}{Value}

\item{strataVars}{Value}
}
\description{
stratification defined by strataVars
}
