% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{mostfreqfunc}
\alias{mostfreqfunc}
\title{Function to find mode of a categorical variable}
\usage{
mostfreqfunc(x)
}
\arguments{
\item{x}{value}
}
\description{
Function to find mode of a categorical variable
}
\keyword{internal}
