% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{norm.mean}
\alias{norm.mean}
\title{Calculate normal mean from lognormal mean and se}
\usage{
norm.mean(x1, x1e)
}
\arguments{
\item{x1}{Value}

\item{x1e}{Value}
}
\description{
Calculate normal mean from lognormal mean and se
}
