% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{norm.se}
\alias{norm.se}
\title{Calculate standard error from lognormal mean and se}
\usage{
norm.se(x1, x1e)
}
\arguments{
\item{x1}{Value}

\item{x1e}{Value}
}
\description{
Calculate standard error from lognormal mean and se
}
