% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BycatchEstimator.R
\docType{data}
\name{obsdatExample}
\alias{obsdatExample}
\title{Observer simple example data set}
\format{
A tibble with 5 columns.
\describe{
\item{EW}{East/West spatial grouping variable}
\item{season}{season temporal grouping variable}
\item{Year}{Year variable}
\item{sampled.sets}{The effort variable e.g., e.g. 1000 hook hours}
\item{Catch}{Bycatch species in numbers caught}
}
}
\source{
Simulated data
}
\usage{
obsdatExample
}
\description{
Example data set for logbook. Can be used to run bycatch estimation procedure along with observer example data set.
}
\keyword{datasets}
