% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{outlierCountFunc}
\alias{outlierCountFunc}
\title{Function to count outliers, defined as more than numSD standard deviations from the mean.}
\usage{
outlierCountFunc(x, numSD = 8)
}
\arguments{
\item{x}{Value}

\item{numSD}{Value}
}
\description{
Function to count outliers, defined as more than numSD standard deviations from the mean.
}
\keyword{internal}
