% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{plot_catch_sampled}
\alias{plot_catch_sampled}
\title{Make a plot of the observed effort and catches}
\usage{
plot_catch_sampled(
  obsdat,
  bin_size = 1,
  effort_col = "Effort",
  catch_col = "Catch"
)
}
\arguments{
\item{obsdat}{observer data}

\item{bin_size}{spatial bin size}

\item{effort_col}{column with effort data}
}
\description{
Make a plot of the observed effort and catches
}
