% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{plot_effort_sampled}
\alias{plot_effort_sampled}
\title{Make plot of effort and sampling distribution}
\usage{
plot_effort_sampled(logdat, obsdat, bin_size = 1, effort_col = "Effort")
}
\arguments{
\item{logdat}{logbook data}

\item{obsdat}{observer data}

\item{bin_size}{bin size for spatial aggregation (in degrees)}

\item{effort_col}{name of effort column}
}
\value{
Makes a plot of the spatial distribution of total effort and sampled effort, with points sized by fraction sampled.
}
\description{
Make plot of effort and sampling distribution
}
