% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{plotCrossVal}
\alias{plotCrossVal}
\title{Function to plot boxplots of RMSE and ME across folds}
\usage{
plotCrossVal(rmse, me, fileName)
}
\arguments{
\item{rmse}{Value}

\item{me}{Value}

\item{fileName}{Value}
}
\description{
Function to plot boxplots of RMSE and ME across folds
}
