% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{plotIndex}
\alias{plotIndex}
\title{Function to plot abundance index plus minus standard error}
\usage{
plotIndex(
  yearpred,
  modType,
  fileName,
  subtext = "",
  indexVarNames,
  allVarNames,
  startYear,
  common,
  run,
  catchType,
  catchUnit
)
}
\arguments{
\item{yearpred}{Value}

\item{modType}{Value}

\item{fileName}{Value}

\item{subtext}{Value}

\item{indexVarNames}{Value}

\item{allVarNames}{Value}

\item{startYear}{Value}

\item{common}{Value}

\item{run}{Value}

\item{catchType}{Value}

\item{catchUnit}{Value}
}
\description{
Function to plot abundance index plus minus standard error
}
\keyword{internal}
