% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{plotSums}
\alias{plotSums}
\title{Function to plot either total positive trips (binomial) or total catch/bycatch (all other models)}
\usage{
plotSums(
  yearpred,
  modType,
  fileName,
  subtext = "",
  allVarNames,
  startYear,
  common,
  run,
  catchType,
  catchUnit,
  VarCalc,
  facetGroups = NULL
)
}
\arguments{
\item{yearpred}{Value}

\item{modType}{Value}

\item{fileName}{Value}

\item{subtext}{Value}

\item{allVarNames}{Value}

\item{startYear}{Value}

\item{common}{Value}

\item{run}{Value}

\item{catchType}{Value}

\item{catchUnit}{Value}
}
\description{
Function to plot either total positive trips (binomial) or total catch/bycatch (all other models)
}
\keyword{internal}
