% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{plotSumsValidate}
\alias{plotSumsValidate}
\title{Function to plot total catch by all models plus a validation number}
\usage{
plotSumsValidate(
  yearpred,
  trueval,
  fileName,
  colName,
  allVarNames,
  startYear,
  common,
  run,
  catchType,
  catchUnit
)
}
\arguments{
\item{yearpred}{Value}

\item{trueval}{Value}

\item{fileName}{Value}

\item{colName}{Value}

\item{allVarNames}{Value}

\item{startYear}{Value}

\item{common}{Value}

\item{run}{Value}

\item{catchType}{Value}

\item{catchUnit}{Value}
}
\description{
Function to plot total catch by all models plus a validation number
}
\keyword{internal}
