% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{ratioVar}
\alias{ratioVar}
\title{Calculate variance of ratio estimator, from data already summarized by strata}
\usage{
ratioVar(x, X, n, N, Rhat, sx2, sy2, sxy)
}
\arguments{
\item{x}{Value}

\item{X}{Value}

\item{n}{Value}

\item{N}{Value}

\item{Rhat}{Value}

\item{sx2}{Value}

\item{sy2}{Value}

\item{sxy}{Value}
}
\description{
Variables are x=sum(obs Effort),X=sum(log Effort), n=observed sample units. N=log sample units, Rhat=mean(obs Catch)/mean(obs Effort), sx2, sy2, and sxy are the observed variance in effort and catch, and covariance
}
\keyword{internal}
