% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{ResidualsFunc}
\alias{ResidualsFunc}
\title{Function plots residuals with both R and Dharma library and calculate residual diagnostics.}
\usage{
ResidualsFunc(modfit1, modType, fileName = NULL, nsim = 250)
}
\arguments{
\item{modfit1}{Value}

\item{modType}{Value}

\item{fileName}{Value}

\item{nsim}{Value}
}
\description{
Function plots residuals with both R and Dharma library and calculate residual diagnostics.
}
\keyword{internal}
