% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{seasonfunc}
\alias{seasonfunc}
\title{Function to convert months into 2, 3 4 or 6 numbered seasons}
\usage{
seasonfunc(month, numseason = 4)
}
\description{
Function to convert months into 2, 3 4 or 6 numbered seasons
}
\keyword{internal}
