% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{simulateGammaDraw}
\alias{simulateGammaDraw}
\title{simulateGammaDraw}
\usage{
simulateGammaDraw(modfit, nObs, b)
}
\arguments{
\item{modfit}{Value}

\item{nObs}{Value}

\item{b}{Value}
}
\description{
simulateGammaDraw
}
\keyword{internal}
