% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{simulateNegBin1Draw}
\alias{simulateNegBin1Draw}
\title{simulateNegBin1Draw}
\usage{
simulateNegBin1Draw(modfit, nObs, b, Effort)
}
\arguments{
\item{modfit}{Value}

\item{nObs}{Value}

\item{b}{Value}

\item{Effort}{Value}
}
\description{
simulateNegBin1Draw
}
