% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{simulateNegBinGam}
\alias{simulateNegBinGam}
\title{Function to simulate DHARMa residuals from a negative binomial GAM or GLM}
\usage{
simulateNegBinGam(modfit, nsims = 250, offsetval = 1)
}
\arguments{
\item{modfit}{Value}

\item{nsims}{Value}

\item{offsetval}{Value}
}
\description{
Function to simulate DHARMa residuals from a negative binomial GAM or GLM
}
