% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{simulateTMBTweedieDraw}
\alias{simulateTMBTweedieDraw}
\title{simulateTMBTweedieDraw}
\usage{
simulateTMBTweedieDraw(modfit, nObs, b, Effort)
}
\arguments{
\item{modfit}{Value}

\item{nObs}{Value}

\item{b}{Value}

\item{Effort}{Value}
}
\description{
simulateTMBTweedieDraw
}
