% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{simulateTweedie}
\alias{simulateTweedie}
\title{Generate simulations to use as input to DHARMa residual calculations for the Tweedie from cpglm.}
\usage{
simulateTweedie(modfit1, nsims)
}
\arguments{
\item{modfit1}{Value}

\item{nsims}{Value}
}
\description{
Generate simulations to use as input to DHARMa residual calculations for the Tweedie from cpglm.
}
\keyword{internal}
