% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{yearfunc}
\alias{yearfunc}
\title{Function to group the specified number of years together, starting from the last year}
\usage{
yearfunc(year, numyears = 1)
}
\description{
Function to group the specified number of years together, starting from the last year
}
\keyword{internal}
