% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchDesign.R
\name{bycatchDesign}
\alias{bycatchDesign}
\title{Bycatch estimation using design-based estimators}
\usage{
bycatchDesign(
  setupObj = setupObj,
  designMethods = "Ratio",
  designVars = "Year",
  designPooling = FALSE,
  poolTypes = NULL,
  pooledVar = NULL,
  adjacentNum = NULL,
  minStrataUnit = 1,
  baseDir = getwd(),
  reportType = "html"
)
}
\arguments{
\item{setupObj}{An object produced by \code{bycatchSetup}.}

\item{designMethods}{Character vector of methods to use for design based estimation. Current options are Ratio and Delta (for a delta-lognormal estimator).}

\item{designVars}{Specify strata that must be included in design based estimates, in order across which data should be pooled. Order of these variables determines order for which pooling will occur.}

\item{designPooling}{TRUE if design-based estimates should be pooled for strata with missing data}

\item{poolTypes}{Type of pooling for each variable in designVars, as a character vector in the same order. Options are "all", "pooledVar" and (currently for year only) "adjacent"}

\item{pooledVar}{Variables to pool over for any variable with pooledVar in the previous line, as a character vector in the same order as designVars. Use NA for variables with other pooling methods.  This can be used to pool (for example) months into seasons when pooling is needed.}

\item{adjacentNum}{Number of adjacent years to include for adjacent pooling, as a numerical vector in the same order as designVars. NA for anything other than year, in the same order as designVars.}

\item{minStrataUnit}{The smallest sample size in the strata defined by designVars that is acceptable, in sample units (e.g. trips); below that pooling will occur.}

\item{baseDir}{Character. A directory to save output. Defaults to current working directory.}

\item{reportType}{Character. Choose type of report to be produced. Options are pdf, html (default) or both.}
}
\description{
Produces estimates of bycatch using design-based ratio estimator and delta estimator, with the option of pooling across stratification variables.
}
\examples{
\dontrun{
library(BycatchEstimator)
#-------------------------------------------------
#Step 1. Run the datasetup function and review data inputs
setupObj<-bycatchSetup(
obsdat = obsdatExample,
logdat = logdatExample,
yearVar = "Year",
obsEffort = "sampled.sets",
logEffort = "sets",
logUnsampledEffort = NULL,
includeObsCatch  = FALSE,
matchColumn = NA,
factorVariables = c("Year","season"),
numericVariables = NA,
EstimateIndex = TRUE,
EstimateBycatch = TRUE,
logNum = NA,
sampleUnit = "trips",
baseDir = getwd(),
runName = "SimulatedExample",
runDescription = "Example with simulated data",
common = "Simulated species",
sp = "Genus species",
obsCatch = "Catch",
catchUnit = "number",
catchType = "dead discard"
)

#-------------
#Step 2. Design-based estimators (with pooling)
designObj <- bycatchDesign(
setupObj = setupObj,
designMethods = c("Ratio", "Delta"),
designVars = c("Year","season"),
designPooling = TRUE,
poolTypes=c("adjacent","all"),
pooledVar=c(NA,NA),
adjacentNum=c(1,NA),
minStrataUnit = 1
)}
}
\keyword{Fitting}
\keyword{functions}
