% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitFunction.R
\name{bycatchFit}
\alias{bycatchFit}
\title{Bycatch model fitting routine}
\usage{
bycatchFit(
  setupObj,
  selectCriteria = "BIC",
  DoCrossValidation = TRUE,
  DredgeCrossValidation = FALSE,
  ResidualTest = TRUE,
  CIval = 0.05,
  VarCalc = "Simulate",
  useParallel = TRUE,
  nSims = 10,
  baseDir = getwd(),
  plotValidation = FALSE,
  trueVals = NULL,
  trueCols = NULL
)
}
\arguments{
\item{setupObj}{An object produced by \code{bycatchSetup}.}

\item{selectCriteria}{Character. Model selection criteria. Options are AICc, AIC and BIC}

\item{DoCrossValidation}{Specify whether to run a 10 fold cross-validation (TRUE or FALSE). This may not work with a small or unbalanced dataset}

\item{DredgeCrossValidation}{DredgeCrossValidation specifies whether to use information criteria to find the best model in cross validation, using the dredge function, or just keep the same model formula. Do not use dredge for very large datasets, as the run will be slow.}

\item{ResidualTest}{Logical. Specify whether to exclude models that fail the DHARMa residuals test.}

\item{CIval}{Specify confidence interval for total bycatch estimates. Should be the alpha level, e.g. 0.05 for 95\%}

\item{VarCalc}{Character. Options are: "Simulate","DeltaMethod", or "None". Variance calculation method. Simulate will not work with a large number of sample units in the logbook data. The delta method for variance calculation is not implemented for the delta-lognormal or delta-gamma methods.}

\item{useParallel}{Logical. Whether to conduct the analysis using parallel processing. Only initialized when more that two cores are available.}

\item{nSims}{Number of simulations used to calculate confidence intervals. Ignored if \code{VarCalc} set to "None"}

\item{baseDir}{Character. A directory to save output. Defaults to current working directory.}

\item{plotValidation}{Logical. Validation. If you have true values of the total bycatch (for example in a simulation study). Make PlotValidation true and fill out the rest of the specification.}

\item{trueVals}{The data set that contains the true simulated total catches by year.}

\item{trueCols}{The column of the true simulated catches that contains true bycatch by year}
}
\description{
Produces model-based estimates of bycatch and annual abundance index, as specified in \code{bycatchSetup}
}
