% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSetupFunction.R
\name{bycatchSetup}
\alias{bycatchSetup}
\title{Bycatch estimation model setup}
\usage{
bycatchSetup(
  modelTry = c("Delta-Lognormal", "Delta-Gamma", "TMBnbinom1", "TMBnbinom2",
    "TMBtweedie"),
  obsdat,
  logdat,
  yearVar,
  obsEffort,
  logEffort,
  logUnsampledEffort = NULL,
  includeObsCatch = FALSE,
  matchColumn = NA,
  factorNames,
  EstimateIndex,
  EstimateBycatch,
  logNum,
  sampleUnit,
  complexModel,
  simpleModel,
  indexModel,
  baseDir = getwd(),
  runName,
  runDescription,
  common,
  sp,
  obsCatch,
  catchUnit,
  catchType
)
}
\arguments{
\item{modelTry}{Specify which observation error models to try. Options are: "Binomial", "Normal","Lognormal", "Delta-Lognormal", "Delta-Gamma", "NegBin" for Negative binomial" using glm.mb in the MASS library, "Tweedie" for Tweedie GLM from the cpglm library, and "TMBnbinom1", "TMBnbinom2", and "TMBtweedie" for negative binomial 1, negative binomial 2 and Tweedie from the GLMMTMB library. Binomial is run automatically as part of the delta models if either of them are selected.}

\item{obsdat}{Observer data set}

\item{logdat}{Logbook data set}

\item{yearVar}{Character. The name of the year variable in \code{obsdat} and \code{logdat}. Both input files must contain the same variable name for year.}

\item{obsEffort}{Character. The name of the effort variable in \code{obsdat}. This variable must have the same effort units as \code{logEffort}}

\item{logEffort}{Character. The name of the effort variable in \code{logdat}. Optional and only used when estimating bycatch. This variable must have the same effort units as \code{obsEffort}}

\item{logUnsampledEffort}{Character. The name of the unsampled effort variable in in \code{logdat}. Optional and used to specify a column for effort that is not sampled, in trips with observers. This can be zero in all cases if observers sample 100\% of effort in sampled trips. Only used when \code{includeObsCatch} is TRUE}

\item{includeObsCatch}{Logical. Set to TRUE if (1) the observed sample units can be matched to the logbook sample units and (2) you want to calculate total bycatch as the observed bycatch plus the predicted unobserved bycatch. This doesn't work with aggregated logbook effort.}

\item{matchColumn}{Character. If \code{includeObsCatch} is TRUE, give the name of the column that matches sample units between the observer and logbook data. Otherwise, this can be NA}

\item{factorNames}{Character vector. Specify which variables should be interpreted as categorical, ensuring imposes factor format on these variables. Variables not in this list will retain their original format. These variables must have identical names and factor levels in \code{obsdat} and \code{logdat}}

\item{EstimateIndex}{Logical. What would you like to estimate? You may calculate either an annual abundance index, or total bycatch, or both.}

\item{EstimateBycatch}{Logical. What would you like to estimate? You may calculate either an annual abundance index, or total bycatch, or both. If you want total bycatch, you must provide logbook data or some other source of total effort to \code{logdat}.}

\item{logNum}{Character vector. The name of the column in \code{logdat} that gives the number of sample units (e.g., trips or sets). If the logbook data is not aggregated (i.e. each row is a sample unit) set value to NA}

\item{sampleUnit}{Character. What is the sample unit in \code{logdat}? e.g. sets or trips.}

\item{complexModel}{Specify as stats::formula. Specify the most complex and simplest model to be considered. The code will find compare all intermediate models using information criteria.}

\item{simpleModel}{Specify as stats::formula.}

\item{indexModel}{Specify as stats::formula. Use indexModel to specify which strata to keep separate in calculating abundance indices.}

\item{baseDir}{Character. A directory to save output. Defaults to current working directory.}

\item{runName}{Characer. Give a name to the run, which will be used to set up a directory for the outputs}

\item{runDescription}{Character. Brief summary of the run, which will be used to set up a directory for the outputs}

\item{common}{Character vector. Provide a common name for the species used in bycatch and index estimation. Can be a vector of names to do multiple species at the same time.}

\item{sp}{Character vector. Provide a scientific name for the species used in bycatch and index estimation. Can be a vector of names to do multiple species at the same time}

\item{obsCatch}{Character vector. The name of the column(s) in \code{obsdat} that contain catch. If it is a vector, order of variable names must follow the same order as names provided in \code{common} and \code{sp}}

\item{catchUnit}{Character vector. Give units of catch (e.g., number) to go in plot labels. Must be a vector of the same length as \code{sp}}

\item{catchType}{Character vector. Give type of catch (e.g., dead discards) to go in plot labels. Must be a vector of the same length as \code{sp}}
}
\description{
Sets global conditions and makes a preliminary data summary.
}
\examples{
\dontrun{
library(BycatchEstimator)
setupObj<-bycatchSetup(
modelTry = c("Delta-Lognormal","Delta-Gamma","TMBnbinom1","TMBnbinom2","TMBtweedie"),
obsdat = obsdatExample,
logdat = logdatExample,
yearVar = "Year",
obsEffort = "sampled.sets",
logEffort = "sets",
logUnsampledEffort = NULL,
includeObsCatch  = FALSE,
matchColumn = NA,
factorNames = c("Year","season"),
EstimateIndex = TRUE,
EstimateBycatch = TRUE,
logNum = NA,
sampleUnit = "trips",
complexModel = formula(y~(Year+season)^2),
simpleModel = formula(y~Year),
indexModel = formula(y~Year),
baseDir = getwd(),
runName = "SimulatedExample",
runDescription = "Example with simulated data",
common = "Simulated species",
sp = "Genus species",
obsCatch = "Catch",
catchUnit = "number",
catchType = "dead discard"
)}
}
\keyword{Fitting}
\keyword{functions}
