% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{findBestModelFunc}
\alias{findBestModelFunc}
\title{Function to find best model by information criteria, by model type}
\usage{
findBestModelFunc(
  obsdatval,
  modType,
  requiredVarNames,
  allVarNames,
  complexModel,
  useParallel,
  selectCriteria,
  varExclude,
  printOutput = FALSE,
  catchType = NULL,
  common = NULL,
  dirname = NULL,
  run = NULL
)
}
\arguments{
\item{obsdatval}{Value}

\item{modType}{Value}

\item{requiredVarNames}{Value}

\item{allVarNames}{Value}

\item{complexModel}{Value}

\item{useParallel}{Value}

\item{selectCriteria}{Value}

\item{varExclude}{Value}

\item{printOutput}{Value}

\item{catchType}{Value}

\item{common}{Value}

\item{dirname}{Value}

\item{run}{Value}
}
\description{
Function to find best model by information criteria, by model type
}
