% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{getDesignEstimates}
\alias{getDesignEstimates}
\title{Function to make design based estimates of bycatch from the
ratio estimator of Pennington Delta estimator, pooling as
needed for strata missing data.
stratification defined by designVars, then aggregated to strataVars}
\usage{
getDesignEstimates(
  obsdatval,
  logdatval,
  strataVars,
  designVars = NULL,
  designPooling,
  minStrataUnit = 1,
  startYear,
  poolingSum = NULL,
  includePool = NULL
)
}
\arguments{
\item{obsdatval}{Value}

\item{logdatval}{Value}

\item{strataVars}{Value}

\item{designVars}{Value}

\item{designPooling}{Value}

\item{minStrataUnit}{Value}

\item{startYear}{Value}

\item{poolingSum}{Value}

\item{includePool}{Value}
}
\description{
Function to make design based estimates of bycatch from the
ratio estimator of Pennington Delta estimator, pooling as
needed for strata missing data.
stratification defined by designVars, then aggregated to strataVars
}
\keyword{internal}
