% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{loadOutputs}
\alias{loadOutputs}
\title{Read in all the R objects created during runs of the bycatchEstimator for further analysis}
\usage{
loadOutputs(
  baseDir = getwd(),
  runName,
  runDate = Sys.Date(),
  loadDesign = TRUE,
  designScenario = NULL,
  loadModel = TRUE,
  modelScenario = NULL
)
}
\arguments{
\item{baseDir}{The base directory for the runs, same as bycatchSetup.}

\item{runName}{The run name, same as bycatchSetup.}

\item{runDate}{The date when the model was run. Defualts to current date, but can be set to read in models previously run.}

\item{loadDesign}{TRUE to read in design-based estimator results.}

\item{loadModel}{TRUE to read in model-based estimator results.}
}
\description{
Read in all the R objects created during runs of the bycatchEstimator for further analysis
}
\keyword{internal}
