% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{makePredictionsDeltaVar}
\alias{makePredictionsDeltaVar}
\title{Generate standard errors and confidence intervals of predictions with delta-method separately by year}
\usage{
makePredictionsDeltaVar(
  modfit1,
  newdat,
  modtype,
  obsdatval,
  includeObsCatch,
  requiredVarNames,
  CIval,
  printOutput = TRUE,
  catchType,
  common,
  dirname,
  run
)
}
\arguments{
\item{modfit1}{Value}

\item{newdat}{Value}

\item{modtype}{Value}

\item{obsdatval}{Value}

\item{includeObsCatch}{Value}

\item{requiredVarNames}{Value}

\item{CIval}{Value}

\item{printOutput}{Value}

\item{catchType}{Value}

\item{common}{Value}

\item{dirname}{Value}

\item{run}{Value}
}
\description{
Generate standard errors and confidence intervals of predictions with delta-method separately by year
}
\keyword{internal}
