% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{makePredictionsNoVar}
\alias{makePredictionsNoVar}
\title{Function to predict total positive trips or total catches with no variance calculation}
\usage{
makePredictionsNoVar(
  modfit1,
  modfit2 = NULL,
  modtype,
  newdat,
  obsdatval = NULL,
  nsims,
  includeObsCatch,
  requiredVarNames,
  printOutput = FALSE,
  catchType = NULL,
  common = NULL,
  dirname = NULL,
  run = NULL
)
}
\arguments{
\item{modfit1}{Value}

\item{modfit2}{Value}

\item{modtype}{Value}

\item{newdat}{Value}

\item{obsdatval}{Value}

\item{nsims}{Value}

\item{includeObsCatch}{Value}

\item{requiredVarNames}{Value}

\item{printOutput}{Value}

\item{catchType}{Value}

\item{common}{Value}

\item{dirname}{Value}

\item{run}{Value}
}
\description{
Function to predict total positive trips or total catches with no variance calculation
}
