% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{makePredictionsNoVar}
\alias{makePredictionsNoVar}
\title{Generate predictions without estimating variance}
\usage{
makePredictionsNoVar(
  modfit1,
  modfit2 = NULL,
  modtype,
  newdat,
  obsdatval = NULL,
  nsims,
  includeObsCatch,
  requiredVarNames,
  printOutput = TRUE,
  catchType,
  common,
  dirname,
  run,
  modelScenario
)
}
\arguments{
\item{modfit1}{Value}

\item{modfit2}{Value}

\item{modtype}{Value}

\item{newdat}{Value}

\item{obsdatval}{Value}

\item{includeObsCatch}{Value}

\item{requiredVarNames}{Value}

\item{printOutput}{Value}

\item{catchType}{Value}

\item{common}{Value}

\item{dirname}{Value}

\item{run}{Value}

\item{modelScenario}{Value}
}
\description{
Generate predictions without estimating variance
}
\keyword{internal}
