% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bycatchFunctions.R
\name{makePredictionsSimVarBig}
\alias{makePredictionsSimVarBig}
\title{Generate standard errors and confidence intervals of predictions from simulation from regression coefficients and their var/covar matrix}
\usage{
makePredictionsSimVarBig(
  modfit1,
  modfit2 = NULL,
  newdat,
  modtype,
  obsdatval,
  includeObsCatch,
  nsim,
  requiredVarNames,
  CIval,
  printOutput = TRUE,
  catchType,
  common,
  shortName,
  dirname,
  run,
  randomEffects,
  randomEffects2,
  modelScenario,
  startYear
)
}
\arguments{
\item{modfit1}{Value}

\item{modfit2}{Value}

\item{newdat}{Value}

\item{modtype}{Value}

\item{obsdatval}{Value}

\item{includeObsCatch}{Value}

\item{nsim}{Value}

\item{requiredVarNames}{Value}

\item{CIval}{Value}

\item{printOutput}{Value}

\item{catchType}{Value}

\item{common}{Value}

\item{shortName}{Value}

\item{dirname}{Value}

\item{run}{Value}

\item{randomEffects}{Value}

\item{randomEffects2}{Value}

\item{modelScenario}{Value}

\item{startYear}{Value}
}
\description{
Generate standard errors and confidence intervals of predictions from simulation from regression coefficients and their var/covar matrix
}
\keyword{internal}
